+++ record/example/lib/audio_recorder.dart
@@ -53,10 +53,7 @@ class _RecorderState extends State<Recorder> with AudioRecorderMixin {
         final devs = await _audioRecorder.listInputDevices();
         debugPrint(devs.toString());
 
-        const config = RecordConfig(
-            encoder: encoder,
-            numChannels: 1,
-            audioInterruption: AudioInterruptionMode.pauseResume);
+        const config = RecordConfig(encoder: encoder, numChannels: 1);
 
         // Record to file
         await recordFile(_audioRecorder, config);