+++ record_ios/ios/record_ios/Sources/record_ios/delegate/RecorderStreamDelegate.swift
@@ -18,24 +18,9 @@ class RecorderStreamDelegate: NSObject, AudioRecordingStreamDelegate {
 
   func start(config: RecordConfig, recordEventHandler: RecordStreamHandler) throws {
     let audioEngine = AVAudioEngine()
-    
-#if os(iOS)
+
     try initAVAudioSession(config: config)
     try setVoiceProcessing(echoCancel: config.echoCancel, autoGain: config.autoGain, audioEngine: audioEngine)
-#else
-    // set input device to the node
-    if let deviceId = config.device?.id,
-       let inputDeviceId = getAudioDeviceIDFromUID(uid: deviceId) {
-      do {
-        try audioEngine.inputNode.auAudioUnit.setDeviceID(inputDeviceId)
-      } catch {
-        throw RecorderError.error(
-          message: "Failed to start recording",
-          details: "Setting input device: \(deviceId) \(error)"
-        )
-      }
-    }
-#endif
     
     let srcFormat = audioEngine.inputNode.inputFormat(forBus: 0)
     
@@ -52,7 +37,7 @@ class RecorderStreamDelegate: NSObject, AudioRecordingStreamDelegate {
         details: "Format is not supported: \(config.sampleRate)Hz - \(config.numChannels) channels."
       )
     }
-    
+
     guard let converter = AVAudioConverter(from: srcFormat, to: dstFormat) else {
       throw RecorderError.error(
         message: "Failed to start recording",
@@ -60,8 +45,12 @@ class RecorderStreamDelegate: NSObject, AudioRecordingStreamDelegate {
       )
     }
     converter.sampleRateConverterQuality = AVAudioQuality.high.rawValue
-    
-    audioEngine.inputNode.installTap(onBus: bus, bufferSize: 2048, format: srcFormat) { (buffer, _) -> Void in
+
+    audioEngine.inputNode.installTap(
+      onBus: bus,
+      bufferSize: AVAudioFrameCount(config.streamBufferSize ?? 1024),
+      format: srcFormat) { (buffer, _) -> Void in
+
       self.stream(
         buffer: buffer,
         dstFormat: dstFormat,
@@ -79,13 +68,11 @@ class RecorderStreamDelegate: NSObject, AudioRecordingStreamDelegate {
   }
   
   func stop(completionHandler: @escaping (String?) -> ()) {
-    #if os(iOS)
     if let audioEngine = audioEngine {
       do {
         try setVoiceProcessing(echoCancel: false, autoGain: false, audioEngine: audioEngine)
       } catch {}
     }
-    #endif
     
     audioEngine?.inputNode.removeTap(onBus: bus)
     audioEngine?.stop()