+++ record_android/CHANGELOG.md
@@ -1,5 +1,6 @@
-## next
+## 1.4.0
 * feat: Add AudioInterruptionMode to `RecordConfig`.
+* feat: Add stream buffer size option.
 * fix: malformed AAC on Android.
 
 ## 1.3.3