+++ record_ios/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0
 * feat: Add AudioInterruptionMode to `RecordConfig`.
+* feat: Add stream buffer size option.
 
 ## 1.0.0
 * chore: Support Swift Package Manager.