+++ record_macos/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.0
+* feat: Add stream buffer size option.
+
 ## 1.0.0
 * chore: Support Swift Package Manager.
 * feat: Enable `echoCancel` & `autoGain` for streaming mode.